# hackaton_back/app.js
const cors = require('cors');
const express = require('express');
const swaggerUi = require('swagger-ui-express');
const YAML = require('yamljs');
const app = express();
const port = 5000;

app.use(cors());

// Load Swagger doc
const swaggerDocument = YAML.load('./openapi.yaml');
app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));

// Dummy data
const communes = [
  {
    nom: "Aiguebelette-le-Lac",
    code: "73001",
    codeDepartement: "73",
    siren: "217300011",
    codeEpci: "247300668",
    codeRegion: "84",
    codesPostaux: ["73610"],
    population: 247
  },
  {
    nom: "Chambéry",
    code: "73065",
    codeDepartement: "73",
    siren: "217300065",
    codeEpci: "247300489",
    codeRegion: "84",
    codesPostaux: ["73000", "73011", "73012"],
    population: 58839
  },
  {
    nom: "Aix-les-Bains",
    code: "73008",
    codeDepartement: "73",
    siren: "217300008",
    codeEpci: "247300522",
    codeRegion: "84",
    codesPostaux: ["73100"],
    population: 30790
  },
  {
    nom: "Albertville",
    code: "73011",
    codeDepartement: "73",
    siren: "217300011",
    codeEpci: "247300546",
    codeRegion: "84",
    codesPostaux: ["73200"],
    population: 19133
  },
  {
    nom: "Saint-Jean-de-Maurienne",
    code: "73248",
    codeDepartement: "73",
    siren: "217302048",
    codeEpci: "247300490",
    codeRegion: "84",
    codesPostaux: ["73300"],
    population: 7915
  }
];

const medecins = [
  {
    nom: "Dr Jean Dupont",
    adresse: "13 rue du grand BAC 73125 Chambéry"
  },
  {
    nom: "Dr Claire Martin",
    adresse: "20 avenue de la Gare 73200 Albertville"
  },
  {
    nom: "Dr Pierre Morel",
    adresse: "5 rue de Genève 73100 Aix-les-Bains"
  },
  {
    nom: "Dr Sophie Lacroix",
    adresse: "42 boulevard de la Colonne 73000 Chambéry"
  },
  {
    nom: "Dr Paul Simon",
    adresse: "8 chemin des Écoles 73300 Saint-Jean-de-Maurienne"
  }
];

app.get('/', (req, res) => {
  res.send('Hello World!');
});

// ---- 1. Liste des communes du département 73
app.get('/departements/73/communes', (req, res) => {
  res.status(200).json(communes);
});

// ---- 2. Liste des médecins du département 73, filtable par codePostal
app.get('/departements/73/medecins', (req, res) => {
  const { codePostal } = req.query;
  if (codePostal) {
    const filtered = medecins.filter(m => m.adresse.includes(codePostal));
    return res.status(200).json(filtered);
  }
  res.status(200).json(medecins);
});

// ---- 3. Endpoint pour la population d'une commune par code INSEE
app.get('/departements/73/communes/:code', (req, res) => {
  const { code } = req.params;
  const commune = communes.find(c => c.code === code);
  if (!commune) return res.status(404).json({ message: "Commune non trouvée" });
  res.json({ population: commune.population });
});

// ---- 4. Recherche des médecins proches des coordonnées (mock)
app.get('/recherche/medecins', (req, res) => {
  const { lat, lon } = req.query;
  // contrôle de la présence et de la validité des coordonnées
  if (!lat || !lon || isNaN(+lat) || isNaN(+lon)) {
    return res.status(400).json({ message: "Coordonnées invalides" });
  }
  // Pour l'exemple, on retourne les 3 premiers médecins
  res.json(medecins.slice(0, 3));
});

app.listen(port, () => {
  console.log(`Example app listening on port ${port}`);
});