# hackaton_back/openapi.yaml
openapi: 3.0.3
info:
  title: API des Départements - Communes et Médecins
  version: 1.0.0
  description: Fournit les communes et les médecins associés à un département donné.

paths:
  /departements/73/communes:
    get:
      summary: Obtenir les communes d'un département
      description: Retourne la liste des communes pour le département spécifié.
      responses:
        '200':
          description: Liste des communes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Commune'
        '400':
          description: Requête invalide
        '404':
          description: Département non trouvé

  /departements/73/medecins:
    get:
      summary: Obtenir les médecins d'un département
      description: Retourne la liste des médecins pour le département spécifié. Peut être filtré par code postal.
      parameters:
        - name: codePostal
          in: query
          required: false
          description: Filtrer les médecins par code postal
          schema:
            type: string
            example: "73125"
      responses:
        '200':
          description: Liste des médecins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Medecin'
        '400':
          description: Requête invalide
        '404':
          description: Données non trouvées
          
  /departements/73/communes/{code}:
    get:
      summary: Obtenir la population d'une commune
      description: Retourne uniquement la population d'une commune donnée dans un département.
      parameters:
        - name: code
          in: path
          required: true
          description: Code INSEE de la commune
          schema:
            type: string
      responses:
        '200':
          description: Population de la commune
          content:
            application/json:
              schema:
                type: object
                properties:
                  population:
                    type: integer
                    example: 10000
        '404':
          description: Commune non trouvée
          
  /recherche/medecins:
    get:
      summary: Rechercher les médecins à proximité de coordonnées
      description: Retourne une liste de médecins les plus proches des coordonnées GPS fournies.
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitude de la position de recherche
          schema:
            type: number
            format: float
            example: 45.769
        - name: lon
          in: query
          required: true
          description: Longitude de la position de recherche
          schema:
            type: number
            format: float
            example: 5.836
      responses:
        '200':
          description: Liste des médecins proches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Medecin'
        '400':
          description: Coordonnées invalides

components:
  schemas:
    Commune:
      type: object
      properties:
        nom:
          type: string
          example: Aiguebelette-le-Lac
        code:
          type: string
          example: "73001"
        codeDepartement:
          type: string
          example: "73"
        siren:
          type: string
          example: "217300011"
        codeEpci:
          type: string
          example: "247300668"
        codeRegion:
          type: string
          example: "84"
        codesPostaux:
          type: array
          items:
            type: string
          example: ["73610"]
        population:
          type: integer
          example: 247

    Medecin:
      type: object
      properties:
        nom:
          type: string
          example: Dr Jean Dupont
        adresse:
          type: string
          example: 13 rue du grand BAC 73125